dev-requirements.txt
nose==1.3.0
Sphinx==1.2b3
numpydoc==0.4
